requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
imagehash
requests
passlib[bcrypt]